backend/pyproject.toml
[tool.poetry]
name = "pixsync-backend"
version = "0.1.0"
description = "Backend for Pixsync using FastAPI with face recognition capabilities"
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
numpy = "^1.26.4"
pandas = "^2.2.2"
scipy = "^1.13.0"
scikit-learn = "^1.4.2"
Pillow = "^10.3.0"
opencv-python = "^4.9.0.80"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
tensorflow = "^2.16.1"
transformers = "^4.41.1"
diffusers = "^0.27.2"
accelerate = "^0.30.1"
matplotlib = "^3.9.0"
requests = "^2.32.2"
python-multipart = "^0.0.9"
retina-face = "^0.0.13"
insightface = "^0.7.3"
hnswlib = "^0.8.0"
openpyxl = "^3.1.2"
faiss-cpu = "1.8.0"
tqdm = "^4.67.1"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
black = "^24.0.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"

[tool.poetry.scripts]
start = "uvicorn src.main:app --host 0.0.0.0 --port 8000 --reload"
dev = "uvicorn src.main:app --host 0.0.0.0 --port 8000 --reload --log-level debug"
face-index = "python -c \"from src.services.face_indexer import FaceIndexer; FaceIndexer('src/photos').build()\""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
